chore(api): Move settings endpoints to their own file and add reset
The reset endpoints are just stubs, but now that we’re adding more things under
the /settings uri we might as well have them somewhere other than endpoints.__init__.
